Create Running Cypher statements
Running Cypher statements
You've been provided with a graph populated with people and companies. The graph schema has been printed for you, showing nodes of type Person and Company, and a relationship named WORKS_FOR between these two node types.

A Cypher statement that finds the (:Person) node with the name property equal to the $name parameter and returns the name property of the company they work for has been provided as cypher.

Execute the Cypher statement on graph to find out where Jo Cornelissen works.

Instructions

Execute the Cypher statement stored in the cypher variable on graph.
Extract the 'company' from each result to return the matched companies.