Implement runnable lambda for reverse function
Runnable lambdas
The RunnableLambda class is a crucial component in a hybrid retrieval chain. Before you use it to execute a vector retriever, let's first practice using it to convert a Python function into an LCEL runnable.

The user-defined reverse() function takes a string and returns it in reverse order. Update the code to return the input string in reverse.

Instructions

Convert the reverse() function into an LCEL runnable.
Call the reverse_chain runnable to reverse the string provided.