Create Assigning additional values to an input
Assigning additional values to an input
While the RunnableLambda function returns a single value, you will also need to preserve the current state. RunnablePassthrough does this by appending new keys to a dictionary holding the user input, so the inputs and results from RunnableLambda can be stored alongside one another.

Complete the code to add an additional output key to the chain's input.

Instructions

Use the correct runnable class to append an output key to the chain's input.
Extract the correct key from the input to multiply by 2 to complete the lambda function.