Create Creating a Ragas evaluation 1# Create an evaluation LLM evaluator_llm = LangchainLLMWrapper(llm)
Creating a Ragas evaluation
To evaluate using ragas, you will need to choose an LLM, create an evaluation dataset, and choose a set of metrics to evaluate against. As you learned in the previous video, two key Graph RAG metrics are context precision and noise sensitivity.

An llm has been defined for you, along with example results (cypher_result) that that demonstrate the differences between plain vector retrieval, text-to-cypher, and hybrid retrieval.

Instructions 1/4

Create a Ragas-compatible LLM from llm, which is an instance of ChatOpenAI from langchain_openai.